Fixed a bunch of g++ warnings
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@14406 01821946-851b-0410-b7f4-f1477c70e57a

+++ network/autodl.cpp
@@ -133,7 +133,7 @@ return m_iTimeout;
 
 bool CDownloadManager::Downloading (uint32_t nPlayer) 
 { 
-if (nPlayer == N_LOCALPLAYER)
+if (nPlayer == (uint32_t) N_LOCALPLAYER)
 	return m_bDownloading [MAX_PLAYERS - 1];
 if (nPlayer >= MAX_PLAYERS) 
 	return false;
